// src/data/schema.ts
import exp from "constants";

export type Item = {
  id: number;
  name: string;
  quantity: number;
}

export type Entity = {
  id: number;
  customer: string;
  items: Item[];
  status: string;
}

export type Film = {
  // title: string;
  // episode_id: number;
  // opening_crawl: string;
  // director: string;
  // producer: string;
  // release_date: string; 
  // species: string[];
  // starships: string[];
  // vehicles: string[];
  // characters: string[];
  // planets: string[];
  // url: string;
  // created: string; 
  // edited: string; 

  id: number;
  customer: string;
  items: Item[];
  status: string;
};

export type Inventory={
  id: number;
  name : string;
  stock: number;
}